Extract reusable 'FeedStoreSpecs' helper methods into shared scope
